Refactor tuple construction for improved readability and consistency
- Updated various instances of tuple construction to use the unpacking operator (*) for clarity.
- Modified shape definitions in multiple files to enhance readability, particularly in grid and field classes.
- Adjusted error messages and assertions to utilize the new tuple formatting.
- Ensured consistent use of unpacking across the codebase for better maintainability.

+++ pde/solvers/base.py
@@ -386,7 +386,7 @@ class SolverBase:
             # retrieve last post_step_data and continue with this
             post_step_data = self.info["post_step_data"]
             # calculate number of steps that lead to an end time closest to t_end
-            steps = max(1, int(round((t_end - t_start) / dt_float)))
+            steps = max(1, round((t_end - t_start) / dt_float))
             # call the stepper with fixed time steps
             t_last = fixed_stepper(state.data, t_start, steps, post_step_data)
             # keep some stats and data
